% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchResults.R
\name{catchActRaw}
\alias{catchActRaw}
\title{catchResults}
\usage{
catchActRaw(result_list)
}
\arguments{
\item{result_list}{The result list from modeling fitting.}

\item{rawData}{Input raw data (quantitative phospho-proteomics / ubiqutin-proteomics data)}

\item{eval_table}{eval_table A tibble or data.frame produced by `catchEvaluations()`.}
}
\description{
Catch Activity and Affinity.
}
