% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JUMPsem.R
\name{JUMPsem}
\alias{JUMPsem}
\title{JUMPsem}
\usage{
JUMPsem(
  input = NULL,
  datatype = NULL,
  organism = NULL,
  enzyme.organism = NULL,
  database = NULL,
  kmo.off = 0,
  mdsite = TRUE,
  enzyList = NULL,
  input.log2.norm = FALSE,
  relative.norm.p = TRUE,
  whole.log2.trans = FALSE,
  whole.proteome = NULL,
  relative.norm.w = TRUE,
  motif = NULL,
  output.folder = getwd()
)
}
\arguments{
\item{input}{Dataframe. Your quantitative ubiquiti-proteomics, phospho-proteomic and acetyl-proteomics data.}

\item{datatype}{Factor. Data type: ubiquitination or phosphoralation or acetylation; Corresponding values are "ubi", "psp" and "ace".}

\item{organism}{Vector. Substrate species. Corresponding values are “human”, “mouse” and “rat”.}

\item{enzyme.organism}{Vector. Enzyme species; Corresponding values are “human”, “mouse” and “rat”.}

\item{database}{Dataframe. Customized database input. Default is NULL using internal database.}

\item{kmo.off}{Numeric. Set up KMO cutoff value 0-1. Default is 0.}

\item{mdsite}{Logical. Mapping with precise phosphorylated modification sites or not. Default is TRUE.}

\item{enzyList}{Vector. Program only calculate the enzyme in the enzyList. Default is to output ALL enzyme activities and affinities.}

\item{input.log2.norm}{Logical. FALSE or TRUE. Need program to do log2 transforming of the input file or not. Default is FALSE.}

\item{relative.norm.p}{Logical. FALSE or TRUE. Need program to do relative normalization of the PTM input file or not. Default is TRUE.}

\item{whole.log2.trans}{Logical. FALSE or TRUE. Need program to do log2 transforming of the whole proteome or not. (Ignore if -whole.proteome is missing). Default is FALSE.}

\item{whole.proteome}{Dataframe. Set up whole proteome used to normalize phosphor-proteome or ubiquitin-proteome whole proteome. Default is NULL.}

\item{relative.norm.w}{Logical. FALSE or TRUE. Need program to do relative normalization of the whole proteomic input file or not. Default is TRUE.}

\item{motif}{Matrix. Added kinase-substrate relationships from motif discovery.}

\item{output.folder}{Character. Character vector of location to save files if desired. Default is current directory.}
}
\value{
List
}
\description{
Infer enzyme activity from phosphoralation, ubiquitination, acetylation
}
\details{
You can use JUMPsem package to infer enzyme activity.
}
\examples{

result <- JUMPsem(input = input_ubi_example,
                  datatype = "ubi",
                  organism = "human",
                  input.log2.norm = T)


result <- JUMPsem(input = input_psp_example,
                  datatype = "psp",
                  organism = "mouse",
                  enzyme.organism = c("human", "mouse", "rat"),
                  input.log2.norm = TRUE,
                  whole.log2.trans = TRUE,
                  motif = motif_example,
                  whole.proteome = wholeProteome_example)


result <- JUMPsem(input = input_ace_example,
                  datatype = "ace",
                  organism = "human",
                  input.log2.norm = FALSE)




}
